+++ src/main/kotlin/icons/RectangleActionsIcons.kt
@@ -8,6 +8,8 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
+@file:Suppress("unused")
+
 package icons
 
 import com.intellij.openapi.util.IconLoader